requirements.txt
python-docx==1.2.0
docx2pdf==0.1.8
PyPDF2==3.0.1
